+++ tests/exporter/graph_ops/test_embed_output_format.py
@@ -3,8 +3,7 @@ import onnx
 from functools import partial
 from onnx import helper
 from onnx import AttributeProto, TensorProto, GraphProto
-from vortex.runtime.onnx.graph_ops.helper import get_Ops
-from vortex.runtime.onnx.helper import get_output_format
+from vortex.runtime.onnx.graph_ops.helper import get_Ops, get_output_format
 
 # shorter version of registry
 from vortex.development import ONNX_GRAPH_OPS as graph_ops
